Update README.md
Towards the end of 2020 and on the onset of 2021 Covid19 took the world by surprise to the extent that it disrupted the livelihood of the human race.  In a journey to the past I sought to analyze the dark part of our life for it is significantly relevant if the future makes us relive a different yet similar experience.